+++ src/features/commands/handlers/behaviors/behaviors.ts
