+++ src/features/commands/handlers/lifecycle/chatSpawn.ts
