+++ src/features/commands/handlers/meta/help.ts
