+++ src/features/index.ts
@@ -1,11 +1,11 @@
-export * from './commands/handlers/help'
-export * from './commands/handlers/chatSpawn'
-export * from './commands/handlers/behaviors'
-export * from './commands/handlers/breakBlock'
+export * from './commands/handlers/meta/help'
+export * from './commands/handlers/lifecycle/chatSpawn'
+export * from './commands/handlers/behaviors/behaviors'
+export * from './commands/handlers/behaviors/breakBlock'
 export * from './commands/handlers/youAreMine'
 export * from './task'
 export * from './gui'
 export * from './killedBySimPlayer'
-export * from './commands/handlers/setting'
-export * from './commands/handlers/showCommandsList'
+export * from './commands/handlers/meta/setting'
+export * from './commands/handlers/meta/showCommandsList'
 export * from './tps/indicator'
